pyproject.toml
[tool.poetry]
name = "twitch-compyle"
version = "0.1.0-alpha.1"
description = "Twitch Compyle is an app that allows you to create Youtube compilations of the latest Twitch clips on your favorite games."
authors = [
    "adesvall <adesvall@users.noreply.github.com>",
    "amaurylrd <amaurylrd@users.noreply.github.com>",
    "thomasabreudias <thomasabreudias@users.noreply.github.com>"
]

repository = "https://github.com/amaurylrd/twitch_compyle"
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
python-dotenv = "^0.21.0"
pymongo = {extras = ["srv"], version = "^4.3.3"}
jsonschema = "^4.0.1"
aenum = "3.1.11"
djangorestframework = "^3.14"
requests = "2.28.1"
moviepy = "1.0.3"
opencv-python = "4.6.0.66"

[tool.poetry.dev-dependencies]
bandit = "^1.7.4"

[tool.poetry.group.lint.dependencies]
black = "^22.10.0"
pylama = {extras = ["all"], version = "^8.4.1"}
isort = "^5.10.1"

[tool.poetry.group.doc.dependencies]
pdoc = "13.1.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.2.0"]

[tool.black]
exclude = '''
(
    /(
        \.git
        |\.tox
        |migrations
        |\.venv
    )/
)
'''
include = '\.pyi?$'
line-length = 120

[tool.isort]
profile = "black"
known_first_party = "compyle"

[tool.pylama]
format = "pycodestyle"
linters = "pycodestyle,pyflakes,mccabe,pylint"
skip = "resources/*,.venv/*,tests/*"
max_line_length = 120

[tool.pylama.linter.pycodestyle]
ignore = "E501,E261,E201,E202,E302,E305,E731,W6,E402,W503,W605"
max_line_length = 120

[tool.pylama.linter.pylint]
disable = "R,E1101,E1130,E1136,E0401,C0103,C0114,C0115,C0116,W0511,W0603,W0612,W0613,W0703,W1514,W1401,W0622"
max_line_length = 120
